+++ src/components/interactive-builder/edit-question.modal.tsx
@@ -454,9 +454,9 @@ const EditQuestionModal: React.FC<EditQuestionModalProps> = ({
                   id="min"
                   labelText="Min"
                   value={min || ''}
-                  invalid={parseInt(min) > parseInt(max)}
+                  invalid={parseFloat(min) > parseFloat(max)}
                   invalidText={
-                    parseInt(min) > parseInt(max) ? t('invalidMinMax', 'Min value cannot be greater than max') : ''
+                    parseFloat(min) > parseFloat(max) ? t('invalidMinMax', 'Min value cannot be greater than max') : ''
                   }
                   onChange={(event: React.ChangeEvent<HTMLInputElement>) => setMin(event.target.value)}
                   required
@@ -465,9 +465,9 @@ const EditQuestionModal: React.FC<EditQuestionModalProps> = ({
                   id="max"
                   labelText="Max"
                   value={max || ''}
-                  invalid={parseInt(min) > parseInt(max)}
+                  invalid={parseFloat(min) > parseFloat(max)}
                   invalidText={
-                    parseInt(min) > parseInt(max) ? t('invalidMinMax', 'Min value cannot be greater than max') : ''
+                    parseFloat(min) > parseFloat(max) ? t('invalidMinMax', 'Min value cannot be greater than max') : ''
                   }
                   onChange={(event: React.ChangeEvent<HTMLInputElement>) => setMax(event.target.value)}
                   required
